// my-code-editor-relooking-logic-last-version/Design/my-code-editor-relooking/src/lsp/types.ts
export interface Position {
  line: number;
  character: number;
}

export interface Range {
  start: Position;
  end: Position;
}
// ...
export interface MarkupContent {
  kind: "plaintext" | "markdown";
  value: string;
}

// InsertTextFormat as const object instead of enum
export const InsertTextFormat = {
  PlainText: 1,
  Snippet: 2,
} as const;
export type InsertTextFormat = typeof InsertTextFormat[keyof typeof InsertTextFormat];

export interface TextEdit {
  range: Range;
  newText: string;
}

// Hover
export interface Hover {
  contents: string | MarkupContent | MarkupContent[];
  range?: Range;
}